refactor

+++ AdfGL/AdfGL_Cad2Dlib/Objects/Rectangle.cs
@@ -9,11 +9,11 @@ namespace AdfGL_Cad2Dlib.Objects
         {
         }
 
-        public float X1 { get; set; }
-        public float Y1 { get; set; }
+        public double X1 { get; set; }
+        public double Y1 { get; set; }
 
-        public float X2 { get; set; }
-        public float Y2 { get; set; }
+        public double X2 { get; set; }
+        public double Y2 { get; set; }
 
         public ELineStyle LineStyle { get; set; } = ELineStyle.Solid;
